Added file for working with files, fixed enum
