+++ .gitignore
@@ -1,0 +1,1 @@
+lab_1/.gitignore
